refactor: fix three abstraction-layer violations from codebase audit
- gitCommands.ts: reuse the shared, VS Code-free isGitRepository probe
  (src/utils/system/isGitRepository.ts) instead of a separately maintained
  sync execaSync implementation, so the extension and desktop hosts can't
  drift apart on worktree/submodule handling. Fold the git-install
  package-manager detection into src/utils/git/gitInstallHelp.ts.
- terminalResultToast: fold the pure terminalResultToast mapper (previously
  src/shared/agent/terminalResultPresentation.ts) into
  src/agent/runtime/terminalResultToast.ts, removing the one remaining
  src/shared -> @agent import that inverted the intended dependency
  direction. Drops the now-empty allowlist entry in
  dependencyDirection.vitest.ts.
- agentCreatorFlow: move the description -> suggested tool groups
  keyword-matching heuristic out of the VS Code UI adapter
  (agentCreatorCommands.ts) into the domain flow, so a future non-VS-Code
  host implementing AgentCreatorUI doesn't have to duplicate it.
