+++ src/test-kernel/architecture/dependencyDirection.vitest.ts
@@ -55,11 +55,12 @@ const AGENT_IMPORT_PATTERNS = [
   /\bimport\s*\(\s*['"]@agent\//,
 ];
 
-const SHARED_AGENT_IMPORT_ALLOWLIST = [
-  // Existing pre-relocation edge. Keep this list fixed unless the edge is
-  // deleted; new shared-to-agent imports would recreate the moved inversion.
-  'src/shared/agent/terminalResultPresentation.ts',
-] as const;
+// The one pre-existing shared-to-agent edge (src/shared/agent/
+// terminalResultPresentation.ts) was deleted by folding its mapper back into
+// `@agent/runtime/terminalResultToast.ts` — the file it always needed
+// `ResultEvent` from. Keep this allowlist empty; a new entry would recreate
+// the inversion.
+const SHARED_AGENT_IMPORT_ALLOWLIST: readonly string[] = [];
 const SHARED_AGENT_IMPORT_ALLOWLIST_SET = new Set<string>(
   SHARED_AGENT_IMPORT_ALLOWLIST,
 );
